Move getrequestfixture into new pytest_compat
- This allows importing the function without pulling Django piece
  before Django has gotten a chance to setup
