+++ pytest_django/pytest_compat.py
@@ -1,0 +1,5 @@
+def getfixturevalue(request, value):
+    if hasattr(request, 'getfixturevalue'):
+        return request.getfixturevalue(value)
+
+    return request.getfuncargvalue(value)
